requirements.txt
flask==3.0.3
flask-cors==4.0.1
numpy==1.26.4
pillow==10.4.0
tensorflow==2.16.1
keras==3.3.3
numpy
gunicorn==21.2.0
requests==2.32.3

